Fix `usize` as `u64`

+++ src/engine.rs
@@ -151,7 +151,7 @@ impl<'a, 'b:'a> ExecutionEngine<'a, 'b> for JitEngine<'b> {
                 MCJMM: ptr::null_mut()
             };
             let size = mem::size_of::<LLVMMCJITCompilerOptions>();
-            let result = engine::LLVMCreateMCJITCompilerForModule(&mut ee, (&*module).into(), &mut options, size as u64, &mut out);
+            let result = engine::LLVMCreateMCJITCompilerForModule(&mut ee, (&*module).into(), &mut options, size, &mut out);
             if result == 0 {
                 Ok(ee.into())
             } else {
